Moving auth routes into api and creating auth module.
